FIX Syntax

+++ mc_server_manager.sh
@@ -12,24 +12,33 @@ MAX_RAM="1024"
 A="-Xmx"
 B="M"
 
+function createdir {
+	if [ ! -e $INSTALL_DIR ]; then
+		mkdir $INSTALL_DIR
+	fi
+}
+
+function downloadjar {
+	if [ ! -e "$INSTALL_DIR/craftbukkit-1.9.jar" ]; then
+		cd $INSTALL_DIR
+		wget "$DW_PATH"
+		cd "../"
+	fi
+}
+
+function addeula 
+{
+	if [ ! -e "$INSTALL_DIR/elua.txt" ]; then
+		echo "eula=true" > "$INSTALL_DIR/eula.txt"
+	fi
+}
+
 function reinstall {
 	rm -r $INSTALL_DIR
 	createdir
 	downloadjar
-}
-
-function createdir {
-	mkdir $INSTALL_DIR
-}
-
-function downloadjar {
-	cd $INSTALL_DIR
-	wget "$DW_PATH"
-	cd "../"
-}
-
-function addeula {
-	echo "eula=true" > "$INSTALL_DIR/eula.txt"
+	addeula
+	echo -e "\e[92mReinstall successful !\e[0m"
 }
 
 function uninstall {
@@ -39,33 +48,23 @@ function uninstall {
 	fi
 }
 
-# CREATE DIR
-if [ ! -e $INSTALL_DIR ]; then
-	createdir
-fi
+function startserver {
+	echo "====================================="
+	read -p "Set the amount of RAM [MB]: " MAX_RAM
+	echo "====================================="
+	cd $INSTALL_DIR
+	java "$A$MAX_RAM$B" -jar craftbukkit-1.9.jar -o true
+}
 
-#IF Craftbukkit.jar not exist download it
-if [ ! -e "$INSTALL_DIR/craftbukkit-1.9.jar" ]; then
-	downloadjar
-fi
-
-#Create elua and set true
-if [ ! -e "$INSTALL_DIR/elua.txt" ]; then
-	addeula
-fi
+createdir
+downloadjar
+addeula
 
 #Start server
 if [ $# -eq 1 ] || [ $# -eq 2 ]; then
 	if [ -e "$INSTALL_DIR/craftbukkit-1.9.jar" ] && [ $1 = "start" ]; then
-
-		echo "====================================="
-		read -p "Set the amount of RAM [MB]: " MAX_RAM
-		echo "====================================="
-
-		cd $INSTALL_DIR
-		java "$A$MAX_RAM$B" -jar craftbukkit-1.9.jar -o true
+		startserver
 	else if [ $1 = "reinstall" ]; then
-		
 		reinstall
 	else if [ $1 = "uninstall" ]; then
 		uninstall
@@ -73,3 +72,4 @@ if [ $# -eq 1 ] || [ $# -eq 2 ]; then
 	fi
 	fi
 fi	
+
